third_party/WebKit/LayoutTests/inspector/elements/styles-1/commit-selector.html

<html>
<head>
<style>
#inspected {
  color: green;
}
</style>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step1);

  function step1() {
    TestRunner.addResult('=== Before selector modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    var section = ElementsTestRunner.firstMatchedStyleSection();
    section.startEditingSelector();
    section._selectorElement.textContent = 'hr, #inspected ';
    ElementsTestRunner.waitForSelectorCommitted(step2);
    section._selectorElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
  }

  function step2() {
    TestRunner.addResult('=== After non-affecting selector modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    var section = ElementsTestRunner.firstMatchedStyleSection();
    section.startEditingSelector();
    section._selectorElement.textContent = '#inspectedChanged';
    ElementsTestRunner.waitForSelectorCommitted(step3);
    section._selectorElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
  }

  function step3() {
    TestRunner.addResult('=== After affecting selector modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that renaming a selector updates element styles. <a href="https://bugs.webkit.org/show_bug.cgi?id=70018">Bug 70018</a>.
</p>

<div id="inspected" style="color: red">Text</div>
<div id="other"></div>

</body>
</html>
