third_party/WebKit/LayoutTests/inspector/elements/styles-1/commit-selector-expected.txt

Tests that renaming a selector updates element styles. Bug 70018.

Text
=== Before selector modification ===
[expanded] 
element.style  { ()
    color: red;

[expanded] 
#inspected  { (commit-selector.html:4 -> commit-selector.html:4:1)
/-- overloaded --/     color: green;

[expanded] 
div  { (user agent stylesheet)
    display: block;

=== After non-affecting selector modification ===
[expanded] 
element.style  { ()
    color: red;

[expanded] 
hr, #inspected  { (commit-selector.html:4 -> commit-selector.html:4:5)
/-- overloaded --/     color: green;

[expanded] 
div  { (user agent stylesheet)
    display: block;

=== After affecting selector modification ===
[expanded] 
element.style  { ()
    color: red;

[expanded] [no-affect] 
#inspectedChanged  { (commit-selector.html:4 -> commit-selector.html:4:1)
    color: green;

[expanded] 
div  { (user agent stylesheet)
    display: block;
